requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-gpu
torch==2.1.0+cu118 --extra-index-url https://download.pytorch.org/whl/cu118
python-dotenv
requests
pydantic
